Dart array tests
