--- Fable.Tests.sln
@@ -1,75 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30114.105
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{B82D35C4-B632-42EE-97AB-7E7895E7C2CB}"
-EndProject
-Project("{F2A71F9B-5D33-465A-A702-920D77279786}") = "Fable.Tests.DllRef", "tests\DllRef\Fable.Tests.DllRef.fsproj", "{361F3B44-F9E0-4BCE-96AF-CB569D7AFE68}"
-EndProject
-Project("{F2A71F9B-5D33-465A-A702-920D77279786}") = "Fable.Tests.External", "tests_external\Fable.Tests.External.fsproj", "{7C51BF90-C95C-45AB-A083-2D9D77BE29A6}"
-EndProject
-Project("{F2A71F9B-5D33-465A-A702-920D77279786}") = "Fable.Tests.Spaces", "tests\Project With Spaces\Fable.Tests.Spaces.fsproj", "{EFEBDC7A-990C-44FC-B561-2C174C300089}"
-EndProject
-Project("{F2A71F9B-5D33-465A-A702-920D77279786}") = "Fable.Tests", "tests\Main\Fable.Tests.fsproj", "{AFC4E849-BC5D-48C4-A307-AE7172997F76}"
-EndProject
-Project("{F2A71F9B-5D33-465A-A702-920D77279786}") = "Fable.Tests.React", "tests\React\Fable.Tests.React.fsproj", "{ACE9CDA6-1106-40DB-8E43-E63295188B84}"
-EndProject
-Project("{F2A71F9B-5D33-465A-A702-920D77279786}") = "Fable.Tests.Adaptive", "tests\Adaptive\Fable.Tests.Adaptive.fsproj", "{DB82C6F2-4A97-4236-B798-8BCC5453A38B}"
-EndProject
-Project("{F2A71F9B-5D33-465A-A702-920D77279786}") = "Compiler", "tests\Compiler\Compiler.fsproj", "{E56E235E-F738-4BEC-A4CE-96FA951A337B}"
-EndProject
-Project("{F2A71F9B-5D33-465A-A702-920D77279786}") = "Fable.Tests.Integration", "tests\Integration\Fable.Tests.Integration.fsproj", "{B914DEDE-6555-46C2-BCDA-C96EE86867E6}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{361F3B44-F9E0-4BCE-96AF-CB569D7AFE68}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{361F3B44-F9E0-4BCE-96AF-CB569D7AFE68}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{361F3B44-F9E0-4BCE-96AF-CB569D7AFE68}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{361F3B44-F9E0-4BCE-96AF-CB569D7AFE68}.Release|Any CPU.Build.0 = Release|Any CPU
-		{7C51BF90-C95C-45AB-A083-2D9D77BE29A6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{7C51BF90-C95C-45AB-A083-2D9D77BE29A6}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{7C51BF90-C95C-45AB-A083-2D9D77BE29A6}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{7C51BF90-C95C-45AB-A083-2D9D77BE29A6}.Release|Any CPU.Build.0 = Release|Any CPU
-		{EFEBDC7A-990C-44FC-B561-2C174C300089}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{EFEBDC7A-990C-44FC-B561-2C174C300089}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{EFEBDC7A-990C-44FC-B561-2C174C300089}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{EFEBDC7A-990C-44FC-B561-2C174C300089}.Release|Any CPU.Build.0 = Release|Any CPU
-		{AFC4E849-BC5D-48C4-A307-AE7172997F76}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{AFC4E849-BC5D-48C4-A307-AE7172997F76}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{AFC4E849-BC5D-48C4-A307-AE7172997F76}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{AFC4E849-BC5D-48C4-A307-AE7172997F76}.Release|Any CPU.Build.0 = Release|Any CPU
-		{ACE9CDA6-1106-40DB-8E43-E63295188B84}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{ACE9CDA6-1106-40DB-8E43-E63295188B84}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{ACE9CDA6-1106-40DB-8E43-E63295188B84}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{ACE9CDA6-1106-40DB-8E43-E63295188B84}.Release|Any CPU.Build.0 = Release|Any CPU
-		{DB82C6F2-4A97-4236-B798-8BCC5453A38B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{DB82C6F2-4A97-4236-B798-8BCC5453A38B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{DB82C6F2-4A97-4236-B798-8BCC5453A38B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{DB82C6F2-4A97-4236-B798-8BCC5453A38B}.Release|Any CPU.Build.0 = Release|Any CPU
-		{E56E235E-F738-4BEC-A4CE-96FA951A337B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E56E235E-F738-4BEC-A4CE-96FA951A337B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E56E235E-F738-4BEC-A4CE-96FA951A337B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E56E235E-F738-4BEC-A4CE-96FA951A337B}.Release|Any CPU.Build.0 = Release|Any CPU
-		{B914DEDE-6555-46C2-BCDA-C96EE86867E6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B914DEDE-6555-46C2-BCDA-C96EE86867E6}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B914DEDE-6555-46C2-BCDA-C96EE86867E6}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B914DEDE-6555-46C2-BCDA-C96EE86867E6}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(NestedProjects) = preSolution
-		{361F3B44-F9E0-4BCE-96AF-CB569D7AFE68} = {B82D35C4-B632-42EE-97AB-7E7895E7C2CB}
-		{EFEBDC7A-990C-44FC-B561-2C174C300089} = {B82D35C4-B632-42EE-97AB-7E7895E7C2CB}
-		{AFC4E849-BC5D-48C4-A307-AE7172997F76} = {B82D35C4-B632-42EE-97AB-7E7895E7C2CB}
-		{ACE9CDA6-1106-40DB-8E43-E63295188B84} = {B82D35C4-B632-42EE-97AB-7E7895E7C2CB}
-		{DB82C6F2-4A97-4236-B798-8BCC5453A38B} = {B82D35C4-B632-42EE-97AB-7E7895E7C2CB}
-		{E56E235E-F738-4BEC-A4CE-96FA951A337B} = {B82D35C4-B632-42EE-97AB-7E7895E7C2CB}
-		{B914DEDE-6555-46C2-BCDA-C96EE86867E6} = {B82D35C4-B632-42EE-97AB-7E7895E7C2CB}
-	EndGlobalSection
-EndGlobal
