+++ Fable.sln
@@ -50,7 +50,11 @@ Project("{F2A71F9B-5D33-465A-A702-920D77279786}") = "Fable.Tests.Rust", "tests\R
 EndProject
 Project("{F2A71F9B-5D33-465A-A702-920D77279786}") = "Fable.Tests", "tests\Python\Fable.Tests.fsproj", "{2C27ECA6-6502-4A47-8D19-6102840B911B}"
 EndProject
-Project("{F2A71F9B-5D33-465A-A702-920D77279786}") = "Fable.Tests.Dart", "tests\Dart\Fable.Tests.Dart.fsproj", "{ABB816A4-6158-43C6-AE3F-17056D013C45}"
+Project("{F2A71F9B-5D33-465A-A702-920D77279786}") = "Fable.Library.Dart", "src\fable-library-dart\Fable.Library.Dart.fsproj", "{505AEE8D-2BB6-48D1-ACD2-FB262DB365C4}"
+EndProject
+Project("{F2A71F9B-5D33-465A-A702-920D77279786}") = "Quicktest.Dart", "src\quicktest-dart\Quicktest.Dart.fsproj", "{2971A69B-4A5B-4841-920F-38A5D11F7881}"
+EndProject
+Project("{F2A71F9B-5D33-465A-A702-920D77279786}") = "Fable.Tests.Dart", "tests\Dart\src\Fable.Tests.Dart.fsproj", "{40C7000A-6BC2-42B7-B7E2-3AAC60DAC06D}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -126,10 +130,18 @@ Global
 		{2C27ECA6-6502-4A47-8D19-6102840B911B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{2C27ECA6-6502-4A47-8D19-6102840B911B}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{2C27ECA6-6502-4A47-8D19-6102840B911B}.Release|Any CPU.Build.0 = Release|Any CPU
-		{ABB816A4-6158-43C6-AE3F-17056D013C45}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{ABB816A4-6158-43C6-AE3F-17056D013C45}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{ABB816A4-6158-43C6-AE3F-17056D013C45}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{ABB816A4-6158-43C6-AE3F-17056D013C45}.Release|Any CPU.Build.0 = Release|Any CPU
+		{505AEE8D-2BB6-48D1-ACD2-FB262DB365C4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{505AEE8D-2BB6-48D1-ACD2-FB262DB365C4}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{505AEE8D-2BB6-48D1-ACD2-FB262DB365C4}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{505AEE8D-2BB6-48D1-ACD2-FB262DB365C4}.Release|Any CPU.Build.0 = Release|Any CPU
+		{2971A69B-4A5B-4841-920F-38A5D11F7881}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{2971A69B-4A5B-4841-920F-38A5D11F7881}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{2971A69B-4A5B-4841-920F-38A5D11F7881}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{2971A69B-4A5B-4841-920F-38A5D11F7881}.Release|Any CPU.Build.0 = Release|Any CPU
+		{40C7000A-6BC2-42B7-B7E2-3AAC60DAC06D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{40C7000A-6BC2-42B7-B7E2-3AAC60DAC06D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{40C7000A-6BC2-42B7-B7E2-3AAC60DAC06D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{40C7000A-6BC2-42B7-B7E2-3AAC60DAC06D}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -153,7 +165,9 @@ Global
 		{475E8FB8-6BC4-4F2B-8081-4D5626181A6A} = {3FE69E51-7170-4B63-A1CD-95F6E1233588}
 		{70D726EE-D492-4561-9C86-9C1CAAAC2ABD} = {CF0A8EC3-841F-4A54-B9FC-8D174CCD4A90}
 		{2C27ECA6-6502-4A47-8D19-6102840B911B} = {CF0A8EC3-841F-4A54-B9FC-8D174CCD4A90}
-		{ABB816A4-6158-43C6-AE3F-17056D013C45} = {CF0A8EC3-841F-4A54-B9FC-8D174CCD4A90}
+		{505AEE8D-2BB6-48D1-ACD2-FB262DB365C4} = {C8CB96CF-68A8-4083-A0F8-319275CF8097}
+		{2971A69B-4A5B-4841-920F-38A5D11F7881} = {C8CB96CF-68A8-4083-A0F8-319275CF8097}
+		{40C7000A-6BC2-42B7-B7E2-3AAC60DAC06D} = {CF0A8EC3-841F-4A54-B9FC-8D174CCD4A90}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {58DF9285-8523-4EAC-B598-BE5B02A76A00}
