--- src/fable-library-dart/Fable.Library.fsproj
@@ -1,41 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <TargetFramework>netstandard2.0</TargetFramework>
-        <DefineConstants>$(DefineConstants);FABLE_COMPILER</DefineConstants>
-        <DefineConstants>$(DefineConstants);FX_NO_BIGINT</DefineConstants>
-    </PropertyGroup>
-
-    <ItemGroup>
-        <Compile Include="../fable-library/Global.fs" />
-<!--         <Compile Include="BigInt/n.fsi" /> -->
-<!--        <Compile Include="BigInt/n.fs" />-->
-<!--         <Compile Include="BigInt/z.fsi" /> -->
-<!--        <Compile Include="BigInt/z.fs" />-->
-<!--        <Compile Include="BigInt.fs" />-->
-<!--        <Compile Include="BigInt/q.fsi"/>-->
-<!--        <Compile Include="BigInt/q.fs"/>-->
-<!--        <Compile Include="SystemException.fs" />-->
-<!--        <Compile Include="System.Text.fs" />-->
-<!--        <Compile Include="System.Collections.Generic.fs" />-->
-<!--        <Compile Include="FSharp.Collections.fs" />-->
-<!--        <Compile Include="FSharp.Core.fs" />-->
-<!--        <Compile Include="Native.fs" />-->
-<!--        <Compile Include="MutableMap.fs" />-->
-<!--        <Compile Include="MutableSet.fs" />-->
-<!--        <Compile Include="Choice.fs" />-->
-<!--        <Compile Include="Array.fs" />-->
-<!--        <Compile Include="List.fs" />-->
-        <Compile Include="../fable-library/Seq.fs" />
-<!--        <Compile Include="Seq2.fs" />-->
-<!--        <Compile Include="Range.fs" />-->
-<!--        <Compile Include="Random.fs" />-->
-<!--        <Compile Include="Set.fs" />-->
-<!--        <Compile Include="Map.fs" />-->
-    </ItemGroup>
-
-    <ItemGroup>
-        <ProjectReference Include="..\Fable.Core\Fable.Core.fsproj" />
-    </ItemGroup>
-
-</Project>
