+++ tests/Dart/.gitignore
@@ -1,1 +1,7 @@
+# Files and directories created by pub.
+.dart_tool/
+.packages
+
+# Conventional directory for build output.
+build/
 *.fs.dart
