+++ tests/Dart/src/Fable.Tests.Dart.fsproj
@@ -3,14 +3,16 @@
     <TargetFramework>net6.0</TargetFramework>
   </PropertyGroup>
   <ItemGroup>
-    <ProjectReference Include="../../src/Fable.Core/Fable.Core.fsproj" />
+    <ProjectReference Include="../../../src/Fable.Core/Fable.Core.fsproj" />
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Tests.Util.fs" />
     <Compile Include="ArithmeticTests.fs" />
+    <Compile Include="ArrayTests.fs" />
     <Compile Include="ComparisonTests.fs" />
     <Compile Include="DateTimeTests.fs" />
     <Compile Include="RegexTests.fs" />
     <Compile Include="UnionTests.fs" />
+    <Content Include="..\main.dart" />
   </ItemGroup>
 </Project>
